to_3d/requirements.txt
matplotlib==3.10.5
numpy==2.2.6
opencv-python==4.12.0.88
Pillow==11.3.0
scipy==1.15.3
transformers==4.55.0
